requirements.txt

MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy>=1.9.2
scipy>=0.15.1
scikit-learn>=0.18
matplotlib>=1.4.3
pandas>=0.19
cufflinks==0.17.3
Pillow==8.1.0
pingouin==0.5.0
plotly==5.4.0
py==1.10.0
seaborn==0.11.1
streamlit>=1.8.1
xlrd<2.0
openpyxl>=3.0.0
chart-studio==1.0.0


